requirements.txt
# Django Backend
Django>=5.2,<6
djangorestframework>=3.15
django-cors-headers>=4.3

# Database
psycopg2-binary>=2.9
dj-database-url>=2.1

# Production Server
gunicorn>=21.0
whitenoise>=6.6

# Development
requests>=2.31
